Extract useExercise
